+++ src/extension.ts
@@ -20,14 +20,22 @@ function debounce(func: (...args: any[]) => void, wait: number) {
 }
 
 export function activate(context: vscode.ExtensionContext) {
+
     console.log('Extension "fanuctpp" is now active!');
 
+    // ------------------INITIAL SETUP-------------------
     // Set line numbers for all currently open documents
     vscode.workspace.textDocuments.forEach(document => {
         if (document.languageId === 'fanuctp_ls') {
             setLineNumbers(document);
         }
     });
+
+    // --------------User Configuration----------------
+    const config = vscode.workspace.getConfiguration('fanuctpp');
+    // Enable auto line edits
+    const autoLineRenum = config.get<boolean>('autoLineRenumber', true);
+    const autoSemi = config.get<boolean>('autoLineRenumber', true);
 
     // Set line numbers for documents that are opened
     const disposeOpen = vscode.workspace.onDidOpenTextDocument(document => {
@@ -41,38 +49,57 @@ export function activate(context: vscode.ExtensionContext) {
         lineNumber = getCurrentLineNumber(event.textEditor);
     });
 
-// Debounced handler for text document changes
-const debouncedOnDidChangeTextDocument = debounce(async (event: vscode.TextDocumentChangeEvent) => {
-    if (isAutoUpd) {
-        console.log('automatic change, skipping handler');
-        return;
-    }
-
-    console.log('manual change, using handler');
-
-    if (event.document.languageId === 'fanuctp_ls') {
-        const newLineCreated = event.contentChanges.some(change => change.text.includes('\n'));
-        const lineDeleted = event.contentChanges.some(change => change.rangeLength > 0 && change.text === '');
-
-        if (newLineCreated || lineDeleted) {
-            await updateLineNumbers(event.document, lineNumber, newLineCreated, lineDeleted);
+    // Debounced handler for text document changes
+    const debouncedOnDidChangeTextDocument = debounce(async (event: vscode.TextDocumentChangeEvent) => {
+        if (isAutoUpd) {
+            console.log('automatic change, skipping handler');
+            return;
         }
+
+        console.log('manual change, using handler');
+
+        if (event.document.languageId === 'fanuctp_ls') {
+            const newLineCreated = event.contentChanges.some(change => change.text.includes('\n'));
+            //const lineDeleted = event.contentChanges.some(change => change.rangeLength > 0 && change.text === '');
+            const lineDeleted = event.contentChanges.some(change => {
+                const startLine = change.range.start.line;
+                const endLine = change.range.end.line;
+                const isWholeLineDeleted = change.range.start.character === 0 && change.range.end.character === 0 && startLine !== endLine;
+                return change.rangeLength > 0 && change.text === '' && (isWholeLineDeleted || startLine !== endLine);
+            });
+
+            if (newLineCreated || lineDeleted) {
+                await updateLineNumbers(event.document, lineNumber, newLineCreated, lineDeleted, autoLineRenum, autoSemi);
+            }
+            // Move cursor to column 8 on new line creation
+            if (newLineCreated) {
+                const editor = vscode.window.activeTextEditor;
+                if (editor) {
+                    const newPosition = new vscode.Position(lineNumber - 1, 7);
+                    editor.selection = new vscode.Selection(newPosition, newPosition);
+                }
+            }
+        }
+    }, 25); // Adjust the debounce delay as needed
+
+    const disposeDebounceChange = vscode.workspace.onDidChangeTextDocument(debouncedOnDidChangeTextDocument);
+
+    // Register the standalone command
+    const disposableCommand = vscode.commands.registerCommand('extension.updateLineNumbers', async () => {
+        const editor = vscode.window.activeTextEditor;
+        if (editor && editor.document.languageId === 'fanuctp_ls') {
+            const lineCreated = false;
+            const lineDeleted = false;
+            await updateLineNumbers(editor.document, lineNumber, lineCreated, lineDeleted, true, false);
+        }
+    });
+
+    // Pushing all event listeners and commands to the context
+    context.subscriptions.push(disposeDebounceChange, disposeOpen, disposeSelection, disposableCommand);
+
     }
-}, 250); // Adjust the debounce delay as needed
 
-const disposeDebounceChange = vscode.workspace.onDidChangeTextDocument(debouncedOnDidChangeTextDocument);
-
-// Register the standalone command
-const disposableCommand = vscode.commands.registerCommand('extension.updateLineNumbers', async () => {
-    const editor = vscode.window.activeTextEditor;
-    if (editor && editor.document.languageId === 'fanuctp_ls') {
-        await updateLineNumbers(editor.document, lineNumber, false, false);
-    }
-});
-
-// Pushing all event listeners and commands to the context
-context.subscriptions.push(disposeDebounceChange, disposeOpen, disposeSelection, disposableCommand);
-}
+// ------------------EVENT HANDLER LOGIC-------------------
 
 // Gets the line number the cursor is currently on
 function getCurrentLineNumber(editor: vscode.TextEditor) {
@@ -85,15 +112,26 @@ function getCurrentLineNumber(editor: vscode.TextEditor) {
         console.log('No active editor');
         return -1;
     }
+
 }
 
 // Updates the line numbers in the document
 // Called on document change in total line numbers
-async function updateLineNumbers(document: vscode.TextDocument, currLine: number, newLineCreated: boolean, lineDeleted: boolean) {
+// Called on command execution
+async function updateLineNumbers(document: vscode.TextDocument, currLine: number, lineCreated: boolean, 
+                                lineDeleted: boolean, autoLineRenum: boolean, autoSemi: boolean) {
+    if (!autoLineRenum) {
+        return;
+    }
+    // Get the file name of the document
     const fileName = path.basename(document.fileName);
-
     if (!(fileName in fileDict)) {
         return;
+    }
+
+    // Check if line numbers need to be set again
+    if (lineCreated || lineDeleted) {
+        await setLineNumbers(document);
     }
 
     // Destructure with default values
@@ -105,28 +143,68 @@ async function updateLineNumbers(document: vscode.TextDocument, currLine: number
 
     // console.log(`updt: Current line number: ${currLine}`);
 
-    const text = document.getText();
-    const lines = text.split('\r\n');
-    const lineNumRegex = /^\s*(\d{1,4}):/;
     const edits: vscode.TextEdit[] = [];
 
-    // Ensure new line is not in the headers
-    if (startLine < currLine && currLine < endLine) {
-        if(newLineCreated) {
+    if(lineCreated || lineDeleted) {
+        const inRange = startLine < currLine && currLine < endLine
+        if (!inRange) {
+            return;
+        }
+
+        const text = document.getText();
+        const lines = text.split('\r\n');
+        const lineNumRegex = /^\s*(\d{1,4}):/;
+        const noSemiNumRegex = /^\s*(\d{1,4}):\s*$/;
+        const twoSemiRegex = /\s*;\s*;\s*$/;
+
+        // LINE CREATED
+        if (lineCreated) {
+            const prevLineMatch = lines[currLine - 2].match(noSemiNumRegex);
+            if (prevLineMatch) {
+                const updatedPrevLine = lines[currLine - 2] + ' ;';
+                lines[currLine - 2] = updatedPrevLine;
+                edits.push(vscode.TextEdit.replace(
+                    new vscode.Range(currLine - 2, 0, currLine - 2, lines[currLine - 2].length),
+                    updatedPrevLine
+                ));
+            }
             const newLineNumber = currLine - startLine; // New Line just created
-            const formattedNewLineNumber = newLineNumber.toString().padStart(4, ' ') + ':  ';
+            let formattedNewLineNumber = newLineNumber.toString().padStart(4, ' ') + ':';
+            if (autoSemi){
+                formattedNewLineNumber = newLineNumber.toString().padStart(4, ' ') + ':   ;';
+            }
             lines[currLine-1] = formattedNewLineNumber;
             edits.push(vscode.TextEdit.replace(new vscode.Range(currLine - 1, 0, currLine - 1, lines[currLine - 1].length), formattedNewLineNumber));
         }
+
+        // LINE DELETED
+        if(lineDeleted) {
+            console.log('updt: Line deleted');     
+            const twoSemiMatch = lines[currLine - 1].match(twoSemiRegex);
+            if (twoSemiMatch) {
+                const updatedPrevLine = lines[currLine - 1].replace(twoSemiRegex, '   ;');
+                lines[currLine - 1] = updatedPrevLine;
+                const lineLength = document.lineAt(currLine - 1).text.length;
+                edits.push(vscode.TextEdit.replace(
+                    new vscode.Range(currLine - 1, 0, currLine - 1, lineLength),
+                    updatedPrevLine
+                ));
+            }
+        }
+
         // Iterate over each line in the document starting at next position
         for (let i = (currLine); i < lines.length && i <= (endLine + 1); i++) {
             const match = lines[i].match(lineNumRegex);
             if (match) {
-                const lineNumber = parseInt(match[1], 10);
-                const nextLineNumber = lineNumber + 1; // Example increment
-                const formattedLineNumber = nextLineNumber.toString().padStart(4, ' '); // Ensure 4 characters
+                lineNumber = parseInt(match[1], 10);
+                if (lineCreated) {
+                    lineNumber = lineNumber + 1;
+                }
+                else {
+                    lineNumber = lineNumber - 1;
+                }
+                const formattedLineNumber = lineNumber.toString().padStart(4, ' '); // Ensure 4 characters
                 const nextLineText = lines[i].replace(lineNumRegex, `${formattedLineNumber}:`);
-                //console.log(`updt: Line ${i + 1} changed from ${lines[i]} to ${newLineText}`);
                 edits.push(vscode.TextEdit.replace(new vscode.Range(i, 0, i, lines[i].length), nextLineText));
             }
         }
@@ -139,16 +217,9 @@ async function updateLineNumbers(document: vscode.TextDocument, currLine: number
     
     isAutoUpd = true;
     console.log('updt: START Applying edits: isAutoUpd = ' + isAutoUpd);
-    vscode.workspace.applyEdit(edit);
+    await vscode.workspace.applyEdit(edit);
     isAutoUpd = false;
     console.log('updt: DONE Applying edits: isAutoUpd = ' + isAutoUpd);
-
-    setLineNumbers(document);
-}
-
-// Gets the line number of a regex match
-function getLineNumberFromPosition(position: number, text: string) {
-    return text.substring(0, position).split('\n').length;
 }
 
 // Constructs the fileDict entry for the document
@@ -159,8 +230,6 @@ async function setLineNumbers(document: vscode.TextDocument) {
     const posEndRegex = /\/POS/;
     const endRegex = /\/END/;
     const headerEndRegex = /\/MN/;
-    const text = document.getText();
-    const lines = text.split('\n');
     let headExists: boolean = false;
     let endExists: boolean = false;
     let posExists: boolean = false;
@@ -168,30 +237,28 @@ async function setLineNumbers(document: vscode.TextDocument) {
     let tpLineStart: number = -1; 
     let tpLineEnd: number = -1;
 
-    const headMatch = headerEndRegex.exec(text);
-    if (headMatch) {
-        tpLineStart = getLineNumberFromPosition(headMatch.index, text);
-        headExists = true;
-    }
+    // Iterate through each line to find matches
+    for (let i = 0; i < document.lineCount; i++) {
+        const line = document.lineAt(i);
 
-    // Find matches for posRegex
-    const posMatch = posEndRegex.exec(text);
-    if (posMatch) {
-        tpLineEnd = getLineNumberFromPosition(posMatch.index, text);
-        posExists = true;
-    }
+        if (headerEndRegex.test(line.text)) {
+            tpLineStart = i + 1; // Line numbers are 1-based
+            headExists = true;
+        }
 
-    // Find matches for endRegex if posRegex didn't match
-    if (!posExists) {
-        const endMatch = endRegex.exec(text);
-        if (endMatch) {
-            tpLineEnd = getLineNumberFromPosition(endMatch.index, text);
+        if (posEndRegex.test(line.text)) {
+            tpLineEnd = i + 1;
+            posExists = true;
+            break; // Stop searching if posEndRegex is found
+        }
+
+        if (!posExists && endRegex.test(line.text)) {
+            tpLineEnd = i + 1;
             endExists = true;
         }
     }
 
-    fileDict[fileName] = [tpLineStart, tpLineEnd, (headExists && (endExists || posExists)), lines.length];
-    
+    fileDict[fileName] = [tpLineStart, tpLineEnd, (headExists && (endExists || posExists)), document.lineCount];
     
     console.log(`set: Setting line numbers for ${fileName}`);
     console.log('set: ' + JSON.stringify(fileDict));